restructure fileformats crate
